Refactor the assets folder
