+++ application/frontend/public/assets/game_assets/assembler.png
