+++ application/frontend/public/assets/game_assets/background.jpg
