+++ application/frontend/public/assets/game_assets/cable.png
