+++ application/frontend/public/assets/game_assets/concrete.png
