+++ application/frontend/public/assets/game_assets/copper_ingot.png
