+++ application/frontend/public/assets/game_assets/copper_ore.png
