+++ application/frontend/public/assets/game_assets/crafter.png
