+++ application/frontend/public/assets/game_assets/iron_ingot.png
