+++ application/frontend/public/assets/game_assets/iron_ore_block.png
