+++ application/frontend/public/assets/game_assets/iron_plate.png
