+++ application/frontend/public/assets/game_assets/mine_sound.mp3
