+++ application/frontend/public/assets/game_assets/miner.png
