+++ application/frontend/public/assets/game_assets/miner1.png
