+++ application/frontend/public/assets/game_assets/miner2.png
