+++ application/frontend/public/assets/game_assets/particle.png
