+++ application/frontend/public/assets/game_assets/rock.png
