+++ application/frontend/public/assets/game_assets/rock_block.png
