+++ application/frontend/public/assets/game_assets/screws.png
