+++ application/frontend/public/assets/game_assets/smelter.png
