+++ application/frontend/public/assets/game_assets/wire.png
