+++ application/frontend/public/assets/page_images/team-members/niko.png
