+++ application/frontend/src/routes/about.tsx
@@ -27,7 +27,7 @@ const teamMembers = [
     role: "Team Lead",
     bio: "",
     hiddenLink: "",
-    image: "assets/images/team-members/katy.png",
+    image: "assets/page_images/team-members/katy.png",
 
   },
   {
@@ -36,42 +36,42 @@ const teamMembers = [
     bio: "Click my name!",
     hiddenLink:
       "https://tracker.gg/valorant/profile/riot/rizzabears%23NA1/overview",
-    image: "assets/images/team-members/arizza.png",
+    image: "assets/page_images/team-members/arizza.png",
   },
   {
     name: "Kullathon “Mos” Sitthisarnwattanachai",
     role: "Git Master",
     bio: "I like Oreos and the United States Permanent Residence card.",
     hiddenLink: "https://www.uscis.gov/green-card",
-    image: "assets/images/team-members/mos.png",
+    image: "assets/page_images/team-members/mos.png",
   },
   {
     name: "Arjun Singh Gill",
     role: "Back-end",
     bio: "",
     hiddenLink: "",
-    image: "assets/images/team-members/arjun.jpg",
+    image: "assets/page_images/team-members/arjun.jpg",
   },
   {
     name: "Matthew Weesner",
     role: "Back-end",
     bio: "",
     hiddenLink: "",
-    image: "assets/images/team-members/matt.png",
+    image: "assets/page_images/team-members/matt.png",
   },
   {
     name: "Niko Galedo",
     role: "Front-end",
     bio: "",
     hiddenLink: "",
-    image: "assets/images/team-members/niko.png",
+    image: "assets/page_images/team-members/niko.png",
   },
   {
     name: "Kevin Lam",
     role: "Front-end",
     bio: "erm what the sigma!",
-    hiddenLink: "assets/images/team-members/amogus.png",
-    image: "assets/images/team-members/kevin.png",
+    hiddenLink: "assets/page_images/team-members/amogus.png",
+    image: "assets/page_images/team-members/kevin.png",
   },
 ]
